+++ tests/testthat/test-concordance.R
@@ -1,10 +1,5 @@
 test_that("concordancer works", {
 
-  enron.sample[1] |>
-    quanteda::tokens() |>
-    concordance(enron.sample[2], enron.sample[3:49], search = "wants to") |>
-    expect_error()
-
   concordance(enron.sample[1], enron.sample[2], enron.sample[3:49],
               search = "wants to", token.type = "word") |>
     expect_snapshot()
@@ -15,5 +10,14 @@ test_that("concordancer works", {
 
   concordance(enron.sample[1], enron.sample[2], search = "want*", token.type = "character") |>
     expect_snapshot()
+
+
+  # testing with sentences
+  enron.sents <- tokens(enron.sample, what = "sentence")
+
+  concordance(enron.sents[1], enron.sents[2], enron.sents[3:49],
+              search = "? _EOS_", token.type = "word") |>
+    expect_snapshot()
+
 
 })